wip(preferences): Connect preferences advances
